set player LEDs from USB port_numbers
For XInput and Switch Pro, this sets the player LEDs according to
the value of usb.core.Device.port_numbers. For Fruit jam, player 1
gets the port by the neopixels, which has the port_numbers value
of `(1,)` (a one element tuple) and the silkscreen label of USB1.
Player two gets port_numbers value `(2,)` for silkscreen label
USB2, which is on the opposite side of the power switch by the
Stemma QT I2C port.