it works! (2 gamepads + asyncio)
This runs two instances of the tiled gamepad visualizer, one for
each of two connected USB controllers. It pretty much works fine,
but I had to set the group scale from 2 down to 1 so everything
will fit on the screen. Now the text is a good size, but the
gamepad sprites are too small. Should look better once I scale up
the sprite bitmap and adjust the anchor coordinates.